+++ packages/event/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.5 (2026-08-29)
+
+### 🧱 Updated Dependencies
+
+- Updated @event-driven-platform/aggregate-reference to 0.1.5
+- Updated @event-driven-platform/tenant-reference to 0.1.5
+- Updated @event-driven-platform/types to 0.1.5
+
 ## 0.1.4 (2026-08-25)
 
 ### 🧱 Updated Dependencies