+++ packages/execution/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.6 (2026-08-29)
+
+### 🧱 Updated Dependencies
+
+- Updated @event-driven-platform/types to 0.1.5
+
 ## 0.0.5 (2026-08-25)
 
 ### 🧱 Updated Dependencies