+++ packages/intent/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.2 (2026-08-29)
+
+### 🧱 Updated Dependencies
+
+- Updated @event-driven-platform/tenant-reference to 0.1.5
+- Updated @event-driven-platform/types to 0.1.5
+
 ## 0.2.1 (2026-08-25)
 
 ### 🧱 Updated Dependencies