+++ packages/read-execution-coordinator/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.0.5 (2026-08-29)
+
+### 🚀 Features
+
+- **read-execution-coordinator:** scope requests by tenant ([d228fc0](https://github.com/sanshan/event-driven-platform/commit/d228fc0))
+
+### 🧱 Updated Dependencies
+
+- Updated @event-driven-platform/query to 0.0.6
+
+### ❤️ Thank You
+
+- Aleksandr Lihih @sanshan
+
 ## 0.0.4 (2026-08-25)
 
 ### 🧱 Updated Dependencies