+++ packages/runner/CHANGELOG.md
@@ -1,3 +1,38 @@
+## 0.0.8 (2026-08-29)
+
+### 🚀 Features
+
+- **observability:** add execution pipeline observability ([#145](https://github.com/sanshan/event-driven-platform/pull/145), [#140](https://github.com/sanshan/event-driven-platform/issues/140), [#141](https://github.com/sanshan/event-driven-platform/issues/141), [#144](https://github.com/sanshan/event-driven-platform/issues/144))
+
+### 🩹 Fixes
+
+- sync TypeScript project references ([52901ab](https://github.com/sanshan/event-driven-platform/commit/52901ab))
+
+### 🧱 Updated Dependencies
+
+- Updated @event-driven-platform/operation-event-envelope-factory to 0.0.8
+- Updated @event-driven-platform/operation-handler-resolver to 0.0.7
+- Updated @event-driven-platform/execution-transaction to 0.0.6
+- Updated @event-driven-platform/aggregate-reference to 0.1.5
+- Updated @event-driven-platform/execution-log-store to 0.0.7
+- Updated @event-driven-platform/operation-handler to 0.0.7
+- Updated @event-driven-platform/operation-result to 0.1.5
+- Updated @event-driven-platform/tenant-reference to 0.1.5
+- Updated @event-driven-platform/execution-log to 0.0.7
+- Updated @event-driven-platform/observability to 0.0.2
+- Updated @event-driven-platform/outbox-store to 0.0.7
+- Updated @event-driven-platform/execution to 0.0.6
+- Updated @event-driven-platform/operation to 0.1.6
+- Updated @event-driven-platform/command to 0.0.7
+- Updated @event-driven-platform/outbox to 0.0.7
+- Updated @event-driven-platform/clock to 0.0.7
+- Updated @event-driven-platform/event to 0.1.5
+- Updated @event-driven-platform/types to 0.1.5
+
+### ❤️ Thank You
+
+- Aleksandr Lihih @sanshan
+
 ## 0.0.7 (2026-08-25)
 
 ### 🧱 Updated Dependencies