+++ packages/subject/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.5 (2026-08-29)
+
+This was a version bump only for @event-driven-platform/subject to align it with other projects, there were no code changes.
+
 ## 0.1.4 (2026-08-25)
 
 This was a version bump only for @event-driven-platform/subject to align it with other projects, there were no code changes.